moved About and Hero to sections directory in Home Page
