+++ .vscode/settings.json
@@ -1,3 +1,10 @@
 {
-  "cSpell.words": ["alecdowning", "Appstore", "embla", "jost", "Partykit"]
+  "cSpell.words": [
+    "alecdowning",
+    "Appstore",
+    "embla",
+    "headshot",
+    "jost",
+    "Partykit"
+  ]
 }
