+++ src/components/index.ts
@@ -1,10 +1,9 @@
-export { About } from "./About";
 export { Contact } from "./Contact/Contact";
 export { Footer } from "./Footer";
 export { Header } from "./Header";
-export { Hero } from "./Hero";
 
 export { ProjectsSection, projects, FullProject } from "./Projects";
 export { Button } from "./ui/button";
+export { StackDisplay } from "./StackDisplay";
 
 export * from "./ui";
