+++ src/pages/Home/sections/About.tsx
@@ -1,10 +1,6 @@
-import headshot from "../assets/headshot_cropped.jpg";
-import { StackDisplay } from "./StackDisplay";
+import { StackDisplay } from "@/components";
+import headshot from "@/assets/headshot_cropped.jpg";
 
-/**
- *
- * @returns {Element}
- */
 export function About() {
   const myTechStack = [
     "TypeScript",
