+++ src/pages/Home/sections/index.ts
@@ -1,0 +1,2 @@
+export { About } from "./About";
+export { Hero } from "./Hero";
